inventory_api/requirements/dev.txt
pytest==8.2.2
pytest-django==4.8.0
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==23.0.0